feat(server)!: Added graceful shutdown
Added a quit channel and a wait group to stop all goroutines and
shutdown the server gracefuly.

BREAKING CHANGE: `handle_connection` expects two additional args:
`wg WaitGroup` and `quitCh chan struct{}`. This is required for the implementation
of graceful shutdown as these params are used for communication with the
main thread.